# backend/docs/swagger.yaml
basePath: /api/todos
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: 这是一个接收 Infisical Webhook 通知并管理 Todo 任务的后端服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Infisical Notification API
  version: "1.0"
paths:
...
  /{id}/complete:
    post:
      consumes:
      - application/json
      description: 将指定 ID 的待办事项标记为已完成
      parameters:
      - description: 待办事项 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回已完成的待办事项
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 待办事项不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 标记待办事项完成
      tags:
      - todos
...
schemes:
- http
- https
securityDefinitions:
  WebhookSecret:
    description: Infisical Webhook 签名验证
    in: header
    name: X-Infisical-Signature
    type: apiKey
swagger: "2.0"